peer set: autocomplete peer IDs
This is just a quick example to remind us how we can/should implement
autocomplete for all the commands. Autocomplete seems not to work at all with
click 7.x (the bullseye packaged version) but if you "pypi-install click" to
upgrade to click 8 and then run this:
    eval "$(_VULA_COMPLETE=bash_source vula)"

...then you will be able to tab complete subcommands as well as the VK argument
to the "peer set" command like this:
    vula peer set <tab>

The above is querying organize over dbus. If you are root, it also works to run
the "peer" subcommand of organize, which will instantiate an organize object
(and read the state file) and call it directly without dbus:
    vula organize peer set <tab>

Next steps for autocompletion:
    - refactor this _ac_get_peer_ids function into something generic to use for
      autocompleting other arguments to other commands
    - make the postinst install something in /etc/bash_completion.d/ ?
    - figure out why completion isn't working with click 7.x and if it can be
      made to (since click 8 won't land in Debian stable for a long time still)