Fixes flake8 F841 unused local variable
The value of c was never used, but it is critical that int(k[-2:]) is
called because this code is used to determine if the key is a chunked
key.

Because it was not obvious on first sight why this call was nessessary
the code received an inline comment.
Also using a more specific exception type should help with clarity.