Implement yamlfile test as unittest
The doctest was based on the .gitlab-ci.yml file each change to that
file would have resulted in a change to this test.

In order to decouple the test from the testfile a unittest using mocking
was introduced.
It was implemented using a unittest because doctests should not contain
all the mocking code.