feat: vula UI implementation
A graphical user interface with python 3.9 and Tkinter.
With this GUI the user is able to easily view and edit common
Vula configurations, such as the Vula peers. It communicates
with vula-organize using D-BUS (Desktop BUS) for Inter-process
communication (IPC) and supports almost the same functionalities
as the currently used CLI.

Co-authored-by: Anna Albrecht <annakatharina.albrecht@students.bfh.ch>
Co-authored-by: Adrian Berger <adrian.berger.3@students.bfh.ch>
Co-authored-by: Sven de Gasparo <sven.degasparo@students.bfh.ch>
Co-authored-by: Julian Haldimann <julian.haldimann@students.bfh.ch>
Co-authored-by: Patrick Hirsbrunner <patrick.hirsbrunner@students.bfh.ch>
Co-authored-by: Janick Stucki <janickmarco.stucki@students.bfh.ch>