fix tests for python 3.10; move mitm to contrib
In Python 3.9 (eg debian bullseye) this TypeError refers to "Unicode-objects"
but in Python 3.10 (eg ubuntu 22.04) the exception says "Strings" instead. This
commit relaxes the test to accept either.

There are still some warnings in 3.10, but with this change pytest at least
passes now.

This commit also moves ./vula/mitm to ./contrib/mitm_tests. The pressing reason
is so that pytest doesn't "collect" it; having it under ./vula made everything
fail under podman due to the fact that pyfiglet isn't in the basic deps image.
I'd rather not install the mitm bits as part of the vula package, so, for now
I'm just moving the mitm directory to contrib and leaving it in a broken state.
I'll try to fix it soon.