Add system tray integration
This commit adds support for the system tray to Vula using the pystray
library. The system tray entry can be started either by running the
command `vula tray` or by launching the Vula Tray desktop file.

The system tray entry displays the status of the different Vula
services, displays the list of peers, and allows the user to run certain
actions such as rediscover, repair, and opening the GUI.

Signed-off-by: Severin Kaderli <severin@kaderli.dev>