(WIP) add debian directory
these six files were initially generated by stdeb, but now that it is
apparently EOL (and no longer works in Python 3.12) we're moving away from
using it.

with this commit, it is possible to build a deb with dpkg-buildpackage -uc -tc
inside of a bullseye container ("make shell" in the podman directory).

note that dpkg-buildpackage puts the deb in ../ which, in our containers, is
part of the ephemeral container's filesystem, so, to preserve it you'll want to
"mv ../*deb ." after building to have it outside the container.

next steps:
* try building on various other debian-based distros
* make the deb Makefile targets use dpkg-buildpackage instead of stdeb
* add native RPM packaging (our RPM targets are also is using stdeb currently)
* remove all traces of stdeb