organize zeroconf instances by interface
previously we were organizing them by subnet, which in retrospect was incorrect
because when we are part of two subnets on the same interface (eg, when doing
both v4 and v6, but also when in more subnets), we want to list them all in the
same descriptor.

with this change, we do that.