podman: install everything with non-stdeb debs
vula, vula_libnss, and highctidh are now installed on bookworm without stdeb,
so "make test" actually works again. (it was currently non-working due to
stdeb-related bitrot)

this also removes some podman support for old distributions, and cleans up
various other cruft.

it also removes "--network none" from the deb target, because "python -m build"
requires internet access. This is unfortunate because it means the
dpkg-installed image can't be rebuilt offline (as it could previously on top of
a deps image).