Peer verification with the REUNION protocol
This is a multicast local-area network peer verification process using
the REUNION protocol.

Example usage to verify `host.local.` and `sectionthirtyone.local.`
requires that on `host.local.` a vula user should run the following
verification command:

  vula verify reunion host.local. --verbose --passphrase x

Respectively and simultaneously on `sectionthirtyone.local.` a vula user
should run the following verification command:

  vula verify reunion sectionthirtyone.local. --verbose --passphrase x

This verification process adds a dependency on the `rendez` python
module that provides the REUNION protocol. The implementation is well
typed and the UDP listener is based on the original UDPListener reunion
multicast.py implementation. The newly updated UDPListeners
implementation has been changed to fit the requirements of vula and
should be moved back into REUNION after there is more certainty about
the correct shape of the API. This implementation has type hints and
resolves a number of outstanding bugs.

This also updates `misc/quick-install` to add additional Debian packages
such as `python-is-python3`, `python3-flask`, and `python3-stem` as well
as using the upstream `ggwave` git repo as they have now merged our
packaging improvements.