Update vula verify reunion
This removes the UDPListeners code which has now been reintegrated with
the rendez library. It updates the cli to use `--multicast-port` rather
than `--port` to reduce UX confusion. It additionally adds a
`--bind-mask` option that by default uses the string `all` to select the
vula `iface_prefix_allowed` preference list as an interface mask list.