podman: make dist=oracular work
I think maybe the reason we need CAP_SYS_ADMIN now is because of
https://github.com/systemd/systemd/issues/34565

And, for some reason systemd in oracular is not running systemd-sysusers to
create our users: "systemctl status systemd-sysusers" says "Create System Users
was skipped because no trigger condition checks were met."

Several people with this error reported "systemctl daemon-reexec" fixed it, but
it does not here. Running systemd-sysusers manually does however, and it seems
harmless to do that in bookworm too, so that is the solution for now.